backend/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
flask-cors==3.0.10
psycopg2-binary==2.9.6
gunicorn==21.2.0
prometheus_client==0.17.1
